requirements.txt
Flask==3.1.0
requests==2.32.3
beautifulsoup4==4.13.3
gunicorn==23.0.0
psycopg[binary]==3.2.3
python-dotenv
SQLAlchemy>=2.0.35
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
